requirements.txt
python-telegram-bot==20.8
pillow==10.2.0
python-dotenv==1.0.1
qrcode==7.4.2